requirements.txt
aiogram>=3.0.0
loguru
apscheduler
python-dotenv
aiosqlite
python-dateutil
psutil
geopy
requests
